moving ModelPage -> LocationPage
