+++ src/screens/LocationPage/Enums/MapFilterEnums.js
